LayoutTests/platform/chromium/accessibility/aria-required-expected.txt

  
This tests that aria-required is a usable attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS succeeded is true
PASS succeeded is true
PASS succeeded is true
PASS successfullyParsed is true

TEST COMPLETE
